+++ src/basic/App.tsx
@@ -2,25 +2,18 @@ import { useState, useCallback, useEffect } from "react";
 import { Coupon, Product } from "../types";
 import Button from "./components/ui/Button";
 import Input from "./components/ui/Input";
-import {
-  CartIcon,
-  ShoppingBagIcon,
-  XIcon,
-  TrashIcon,
-  PlusIcon,
-} from "./components/icons";
+import { CartIcon, ShoppingBagIcon } from "./components/icons";
 import ProductList from "./components/product/ProductList";
 import CartList from "./components/cart/CartList";
 import useNotification from "./utils/hooks/useNotification";
 import NotificationToast from "./components/ui/NotificationToast";
-import { isValidNumericInput } from "./utils/validators";
 import { useDebounce } from "./utils/hooks/useDebounce";
 import { formatters } from "./utils/formatters";
-import { parsers } from "./utils/parsers";
 import cartService from "./services/cart";
 import { useCart } from "./hooks/useCart";
 import { useCoupon } from "./hooks/useCoupon";
 import { useProduct } from "./hooks/useProduct";
+import AdminPage from "./pages/AdminPage";
 
 export interface ProductWithUI extends Product {
   description?: string;
@@ -66,59 +59,11 @@ const App = () => {
   // 알람 기능
   const notification = useNotification();
 
-  // 탭
-  const [activeTab, setActiveTab] = useState<"products" | "coupons">(
-    "products"
-  );
-
-  // 쿠폰 표시 여부 (activeTab === 'coupons')
-  const [showCouponForm, setShowCouponForm] = useState(false);
-
-  // 상품 추가 폼 (activeTab ==== 'product')
-  const [showProductForm, setShowProductForm] = useState(false);
-
   // 검색어
   const [searchTerm, setSearchTerm] = useState("");
 
   // 디바운스된 검색어
   const debouncedSearchTerm = useDebounce(searchTerm);
-
-  // Admin
-  const [editingProduct, setEditingProduct] = useState<string | null>(null);
-
-  // 상품 추가 폼 정보
-  const [productForm, setProductForm] = useState({
-    name: "",
-    price: 0,
-    stock: 0,
-    description: "",
-    discounts: [] as Array<{ quantity: number; rate: number }>,
-  });
-
-  // 쿠폰 추가 폼 정보
-  const [couponForm, setCouponForm] = useState({
-    name: "",
-    code: "",
-    discountType: "amount" as "amount" | "percentage",
-    discountValue: 0,
-  });
-
-  // [ui] 상품 가격 포맷팅
-  const formatPrice = useCallback(
-    (price: number, productId?: string) => {
-      if (productId) {
-        const product = products.find((p) => p.id === productId);
-        if (product) {
-          const { isOutOfStock } = getStockStatus(product, cart);
-          if (isOutOfStock) {
-            return "SOLD OUT";
-          }
-        }
-      }
-      return formatters.price(price, !isAdmin);
-    },
-    [products, cart, getStockStatus, isAdmin]
-  );
 
   // [ui] 장바구니 총 상품 수 계산
   const [totalItemCount, setTotalItemCount] = useState(0);
@@ -245,124 +190,11 @@ const App = () => {
     [deleteCoupon, notification.add]
   );
 
-  // [product] 상품 추가/수정 폼 제출 처리
-  const handleProductSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    // editingProduct : new - 새 상품 추가 / edit - 상품 수정
-    // 상품 수정 상태
-    if (editingProduct && editingProduct !== "new") {
-      // [product] 상품 정보 수정 처리
-      handleUpdateProduct(editingProduct, productForm);
-
-      // 수정모드 취소
-      setEditingProduct(null);
-    } else {
-      // 상품 추가 상태
-      handleAddProduct({
-        ...productForm,
-        discounts: productForm.discounts,
-      });
-    }
-    // 상품 추가/수정 폼 초기화
-    setProductForm({
-      name: "",
-      price: 0,
-      stock: 0,
-      description: "",
-      discounts: [],
-    });
-
-    // 수정모드 취소
-    setEditingProduct(null);
-
-    // 상품 추가/수정
-    setShowProductForm(false);
-  };
-
-  // [coupon] 쿠폰 생성 폼 제출 처리
-  const handleCouponSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    // 쿠폰추가
-    handleAddCoupon(couponForm);
-    // 쿠폰 폼 초기화
-    setCouponForm(initializeCoupon());
-    setShowCouponForm(false);
-  };
-
-  // [product] 상품 수정 모드 시작
-  const startEditProduct = (product: ProductWithUI) => {
-    // 수정할 특정 상품 지정
-    setEditingProduct(product.id);
-
-    // 수정할 특정폼에 대한 정보를 가져와서 세팅
-    setProductForm({
-      name: product.name,
-      price: product.price,
-      stock: product.stock,
-      description: product.description || "",
-      discounts: product.discounts || [],
-    });
-    setShowProductForm(true);
-  };
-
   // [cart] 장바구니 총액 계산
   const totals = cartService.calculateCartTotal({
     cart,
     selectedCoupon,
   });
-
-  const handleStockChange = useCallback(
-    (value: string) => {
-      const validation = validateProductStock(value);
-
-      if (validation.isValid) {
-        setProductForm({
-          ...productForm,
-          stock: validation.value,
-        });
-      } else if (validation.message) {
-        notification.add(validation.message, "error");
-        setProductForm({
-          ...productForm,
-          stock: validation.value,
-        });
-      }
-    },
-    [productForm, validateProductStock, notification.add]
-  );
-
-  const handleCouponChange = useCallback(
-    (value: string) => {
-      const discountValue = parsers.safeParseInt(value);
-
-      const validation = validateCoupon(couponForm.discountType, discountValue);
-
-      if (!validation.isValid) {
-        notification.add(validation.message, "error");
-      }
-      setCouponForm({
-        ...couponForm,
-        discountValue: validation.value,
-      });
-    },
-    [couponForm, validateCoupon, notification.add]
-  );
-
-  const handleProductPriceChange = useCallback(
-    (value: string) => {
-      const validation = validateProductPrice(value);
-
-      if (!validation.isValid) {
-        notification.add(validation.message, "error");
-      }
-      setProductForm({
-        ...productForm,
-        price: validation.value,
-      });
-    },
-    [productForm, validateProductPrice, notification.add]
-  );
 
   return (
     <div className="min-h-screen bg-gray-50">
@@ -409,497 +241,16 @@ const App = () => {
 
       <main className="max-w-7xl mx-auto px-4 py-8">
         {isAdmin ? (
-          <div className="max-w-6xl mx-auto">
-            <div className="mb-8">
-              <h1 className="text-2xl font-bold text-gray-900">
-                관리자 대시보드
-              </h1>
-              <p className="text-gray-600 mt-1">
-                상품과 쿠폰을 관리할 수 있습니다
-              </p>
-            </div>
-            <div className="border-b border-gray-200 mb-6">
-              <nav className="-mb-px flex space-x-8">
-                <button
-                  onClick={() => setActiveTab("products")}
-                  className={`py-2 px-1 border-b-2 font-medium text-sm transition-colors ${
-                    activeTab === "products"
-                      ? "border-gray-900 text-gray-900"
-                      : "border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300"
-                  }`}
-                >
-                  상품 관리
-                </button>
-                <button
-                  onClick={() => setActiveTab("coupons")}
-                  className={`py-2 px-1 border-b-2 font-medium text-sm transition-colors ${
-                    activeTab === "coupons"
-                      ? "border-gray-900 text-gray-900"
-                      : "border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300"
-                  }`}
-                >
-                  쿠폰 관리
-                </button>
-              </nav>
-            </div>
-
-            {activeTab === "products" ? (
-              <section className="bg-white rounded-lg border border-gray-200">
-                <div className="p-6 border-b border-gray-200">
-                  <div className="flex justify-between items-center">
-                    <h2 className="text-lg font-semibold">상품 목록</h2>
-                    <Button
-                      onClick={() => {
-                        // 상품 목록에서 새 상품을 추가
-                        setEditingProduct("new");
-                        setProductForm({
-                          name: "",
-                          price: 0,
-                          stock: 0,
-                          description: "",
-                          discounts: [],
-                        });
-                        setShowProductForm(true); // 상품 정보 폼 표시
-                      }}
-                      variant="black"
-                      size="small"
-                    >
-                      새 상품 추가
-                    </Button>
-                  </div>
-                </div>
-
-                <div className="overflow-x-auto">
-                  <table className="w-full">
-                    <thead className="bg-gray-50 border-b border-gray-200">
-                      <tr>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          상품명
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          가격
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          재고
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          설명
-                        </th>
-                        <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          작업
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="bg-white divide-y divide-gray-200">
-                      {(activeTab === "products" ? products : products).map(
-                        (product) => (
-                          <tr key={product.id} className="hover:bg-gray-50">
-                            <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                              {product.name}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                              {formatPrice(product.price, product.id)}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                              <span
-                                className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                                  product.stock > 10
-                                    ? "bg-green-100 text-green-800"
-                                    : product.stock > 0
-                                    ? "bg-yellow-100 text-yellow-800"
-                                    : "bg-red-100 text-red-800"
-                                }`}
-                              >
-                                {product.stock}개
-                              </span>
-                            </td>
-                            <td className="px-6 py-4 text-sm text-gray-500 max-w-xs truncate">
-                              {product.description || "-"}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                              <Button
-                                onClick={() => startEditProduct(product)}
-                                variant="ghost"
-                                size="small"
-                                className="text-indigo-600 hover:text-indigo-900 mr-3"
-                              >
-                                수정
-                              </Button>
-                              <Button
-                                onClick={() => handleDeleteProduct(product.id)}
-                                variant="ghost"
-                                size="small"
-                                className="text-red-600 hover:text-red-900"
-                              >
-                                삭제
-                              </Button>
-                            </td>
-                          </tr>
-                        )
-                      )}
-                    </tbody>
-                  </table>
-                </div>
-                {showProductForm && (
-                  <div className="p-6 border-t border-gray-200 bg-gray-50">
-                    <form onSubmit={handleProductSubmit} className="space-y-4">
-                      <h3 className="text-lg font-medium text-gray-900">
-                        {editingProduct === "new"
-                          ? "새 상품 추가"
-                          : "상품 수정"}
-                      </h3>
-                      <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-                        <div>
-                          <Input
-                            label="상품명"
-                            type="text"
-                            value={productForm.name}
-                            onChange={(e) =>
-                              setProductForm({
-                                ...productForm,
-                                name: e.target.value,
-                              })
-                            }
-                            required
-                          />
-                        </div>
-                        <div>
-                          <Input
-                            label="설명"
-                            type="text"
-                            value={productForm.description}
-                            onChange={(e) =>
-                              setProductForm({
-                                ...productForm,
-                                description: e.target.value,
-                              })
-                            }
-                          />
-                        </div>
-                        <div>
-                          <Input
-                            label="가격"
-                            type="text"
-                            value={
-                              productForm.price === 0 ? "" : productForm.price
-                            }
-                            onChange={(e) => {
-                              const value = e.target.value;
-                              if (isValidNumericInput(value)) {
-                                setProductForm({
-                                  ...productForm,
-                                  price: parsers.parseNumericInput(value),
-                                });
-                              }
-                            }}
-                            onBlur={(e) =>
-                              handleProductPriceChange(e.target.value)
-                            }
-                            placeholder="숫자만 입력"
-                            required
-                          />
-                        </div>
-                        <div>
-                          <Input
-                            label="재고"
-                            type="text"
-                            value={
-                              productForm.stock === 0 ? "" : productForm.stock
-                            }
-                            onChange={(e) => {
-                              const value = e.target.value;
-                              if (isValidNumericInput(value)) {
-                                setProductForm({
-                                  ...productForm,
-                                  stock: parsers.parseNumericInput(value),
-                                });
-                              }
-                            }}
-                            onBlur={(e) => {
-                              handleStockChange(e.target.value);
-                            }}
-                            placeholder="숫자만 입력"
-                            required
-                          />
-                        </div>
-                      </div>
-                      <div className="mt-4">
-                        <label className="block text-sm font-medium text-gray-700 mb-2">
-                          할인 정책
-                        </label>
-                        <div className="space-y-2">
-                          {productForm.discounts.map((discount, index) => (
-                            <div
-                              key={index}
-                              className="flex items-center gap-2 bg-gray-50 p-2 rounded"
-                            >
-                              <Input
-                                type="number"
-                                value={discount.quantity}
-                                onChange={(e) => {
-                                  const newDiscounts = [
-                                    ...productForm.discounts,
-                                  ];
-                                  newDiscounts[index].quantity =
-                                    parsers.safeParseInt(e.target.value);
-                                  setProductForm({
-                                    ...productForm,
-                                    discounts: newDiscounts,
-                                  });
-                                }}
-                                className="w-20 px-2 py-1"
-                                min="1"
-                                placeholder="수량"
-                              />
-                              <span className="text-sm">개 이상 구매 시</span>
-                              <Input
-                                type="number"
-                                value={discount.rate * 100}
-                                onChange={(e) => {
-                                  const newDiscounts = [
-                                    ...productForm.discounts,
-                                  ];
-                                  newDiscounts[index].rate =
-                                    parsers.parseDiscountRate(e.target.value);
-                                  setProductForm({
-                                    ...productForm,
-                                    discounts: newDiscounts,
-                                  });
-                                }}
-                                className="w-16 px-2 py-1"
-                                min="0"
-                                max="100"
-                                placeholder="%"
-                              />
-                              <span className="text-sm">% 할인</span>
-                              <Button
-                                type="button"
-                                onClick={() => {
-                                  const newDiscounts =
-                                    productForm.discounts.filter(
-                                      (_, i) => i !== index
-                                    );
-                                  setProductForm({
-                                    ...productForm,
-                                    discounts: newDiscounts,
-                                  });
-                                }}
-                                variant="ghost"
-                                size="small"
-                                className="text-red-600 hover:text-red-800 p-1 flex items-center justify-center"
-                              >
-                                <XIcon className="w-4 h-4" />
-                              </Button>
-                            </div>
-                          ))}
-                          <button
-                            type="button"
-                            onClick={() => {
-                              setProductForm({
-                                ...productForm,
-                                discounts: [
-                                  ...productForm.discounts,
-                                  { quantity: 10, rate: 0.1 },
-                                ],
-                              });
-                            }}
-                            className="text-sm text-indigo-600 hover:text-indigo-800"
-                          >
-                            + 할인 추가
-                          </button>
-                        </div>
-                      </div>
-
-                      <div className="flex justify-end gap-3">
-                        <Button
-                          type="button"
-                          onClick={() => {
-                            setEditingProduct(null);
-                            setProductForm({
-                              name: "",
-                              price: 0,
-                              stock: 0,
-                              description: "",
-                              discounts: [],
-                            });
-                            setShowProductForm(false);
-                          }}
-                          variant="outline"
-                          size="small"
-                        >
-                          취소
-                        </Button>
-                        <Button type="submit" variant="indigo" size="medium">
-                          {editingProduct === "new" ? "추가" : "수정"}
-                        </Button>
-                      </div>
-                    </form>
-                  </div>
-                )}
-              </section>
-            ) : (
-              <section className="bg-white rounded-lg border border-gray-200">
-                <div className="p-6 border-b border-gray-200">
-                  <h2 className="text-lg font-semibold">쿠폰 관리</h2>
-                </div>
-                <div className="p-6">
-                  <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
-                    {coupons.map((coupon) => (
-                      <div
-                        key={coupon.code}
-                        className="relative bg-gradient-to-r from-indigo-50 to-purple-50 rounded-lg p-4 border border-indigo-200"
-                      >
-                        <div className="flex justify-between items-start">
-                          <div className="flex-1">
-                            <h3 className="font-semibold text-gray-900">
-                              {coupon.name}
-                            </h3>
-                            <p className="text-sm text-gray-600 mt-1 font-mono">
-                              {coupon.code}
-                            </p>
-                            <div className="mt-2">
-                              <span className="inline-flex items-center px-3 py-1 rounded-full text-sm font-medium bg-white text-indigo-700">
-                                {coupon.discountType === "amount"
-                                  ? `${formatters.price(
-                                      coupon.discountValue,
-                                      false
-                                    )} 할인`
-                                  : `${coupon.discountValue}% 할인`}
-                              </span>
-                            </div>
-                          </div>
-                          <Button
-                            onClick={() => handleDeleteCoupon(coupon.code)}
-                            variant="ghost"
-                            size="small"
-                            className="text-gray-400 hover:text-red-600 transition-colors p-1 flex items-center justify-center"
-                          >
-                            <TrashIcon className="w-5 h-5" />
-                          </Button>
-                        </div>
-                      </div>
-                    ))}
-
-                    <div className="border-2 border-dashed border-gray-300 rounded-lg p-4 flex items-center justify-center hover:border-gray-400 transition-colors">
-                      <Button
-                        onClick={() => setShowCouponForm(!showCouponForm)}
-                        variant="ghost"
-                        className="text-gray-400 hover:text-gray-600 flex flex-col items-center justify-center h-full w-full"
-                      >
-                        <PlusIcon className="w-8 h-8" />
-                        <p className="mt-2 text-sm font-medium">새 쿠폰 추가</p>
-                      </Button>
-                    </div>
-                  </div>
-
-                  {showCouponForm && (
-                    <div className="mt-6 p-4 bg-gray-50 rounded-lg">
-                      <form onSubmit={handleCouponSubmit} className="space-y-4">
-                        <h3 className="text-md font-medium text-gray-900">
-                          새 쿠폰 생성
-                        </h3>
-                        <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
-                          <div>
-                            <Input
-                              label="쿠폰명"
-                              type="text"
-                              value={couponForm.name}
-                              onChange={(e) =>
-                                setCouponForm({
-                                  ...couponForm,
-                                  name: e.target.value,
-                                })
-                              }
-                              className="text-sm"
-                              placeholder="신규 가입 쿠폰"
-                              required
-                            />
-                          </div>
-                          <div>
-                            <Input
-                              label="쿠폰 코드"
-                              type="text"
-                              value={couponForm.code}
-                              onChange={(e) =>
-                                setCouponForm({
-                                  ...couponForm,
-                                  code: formatCouponCode(e.target.value),
-                                })
-                              }
-                              className="text-sm font-mono"
-                              placeholder="WELCOME2024"
-                              required
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-sm font-medium text-gray-700 mb-1">
-                              할인 타입
-                            </label>
-                            <select
-                              value={couponForm.discountType}
-                              onChange={(e) =>
-                                setCouponForm({
-                                  ...couponForm,
-                                  discountType: e.target.value as
-                                    | "amount"
-                                    | "percentage",
-                                })
-                              }
-                              className="w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border text-sm"
-                            >
-                              <option value="amount">정액 할인</option>
-                              <option value="percentage">정률 할인</option>
-                            </select>
-                          </div>
-                          <div>
-                            <Input
-                              label={
-                                couponForm.discountType === "amount"
-                                  ? "할인 금액"
-                                  : "할인율(%)"
-                              }
-                              type="text"
-                              value={
-                                couponForm.discountValue === 0
-                                  ? ""
-                                  : couponForm.discountValue
-                              }
-                              onChange={(e) => {
-                                handleCouponChange(e.target.value);
-                              }}
-                              onBlur={(e) => {
-                                handleCouponChange(e.target.value);
-                              }}
-                              className="text-sm"
-                              placeholder={
-                                couponForm.discountType === "amount"
-                                  ? "5000"
-                                  : "10"
-                              }
-                              required
-                            />
-                          </div>
-                        </div>
-                        <div className="flex justify-end gap-3">
-                          <Button
-                            type="button"
-                            onClick={() => setShowCouponForm(false)}
-                            variant="outline"
-                            size="small"
-                          >
-                            취소
-                          </Button>
-                          <Button type="submit" variant="indigo" size="medium">
-                            쿠폰 생성
-                          </Button>
-                        </div>
-                      </form>
-                    </div>
-                  )}
-                </div>
-              </section>
-            )}
-          </div>
+          <AdminPage
+            isAdmin={isAdmin}
+            products={products}
+            coupons={coupons}
+            addProduct={handleAddProduct}
+            updateProduct={handleUpdateProduct}
+            deleteProduct={handleDeleteProduct}
+            addCoupon={handleAddCoupon}
+            deleteCoupon={handleDeleteCoupon}
+          />
         ) : (
           <div className="grid grid-cols-1 lg:grid-cols-4 gap-6">
             <div className="lg:col-span-3">
